+++ core/class/gsl.class.php
@@ -22,6 +22,7 @@ require_once __DIR__ . '/../../../../core/php/core.inc.php';
 class gsl extends eqLogic {
 	/*     * *************************Attributs****************************** */
 	public static $_widgetPossibility = array('custom' => true);
+  	public static $_cookiePath = __DIR__ . '/../config/cookies.txt';
 	/*     * ***********************Methode static*************************** */
 
 	public static function distance($_a, $_b) {
@@ -40,9 +41,10 @@ class gsl extends eqLogic {
 	}
 
 	public static function google_callLocationUrl() {
+		log::add('gsl', 'debug', __('google_callLocationUrl ', __FILE__));
 		$ch = curl_init('https://www.google.com/maps/preview/locationsharing/read?authuser=0&pb=');
-		curl_setopt($ch, CURLOPT_COOKIEJAR, jeedom::getTmpFolder('gsl') . '/cookies.txt');
-		curl_setopt($ch, CURLOPT_COOKIEFILE, jeedom::getTmpFolder('gsl') . '/cookies.txt');
+		curl_setopt($ch, CURLOPT_COOKIEJAR, self::$_cookiePath);
+		curl_setopt($ch, CURLOPT_COOKIEFILE,self::$_cookiePath);
 		curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
 		curl_setopt($ch, CURLOPT_TIMEOUT, 3);
 		curl_setopt($ch, CURLOPT_HEADER, 1);
@@ -59,6 +61,8 @@ class gsl extends eqLogic {
 			throw new Exception(__('Erreur données de localisation n\'est pas un json valide : ', __FILE__) . $result);
 		}
 		$result = json_decode($result, true);
+      	log::add('gsl', 'debug', __('Location data : Connection réussie, reponse : ', __FILE__) .json_encode($result));
+		
 		if (!isset($result[0])) {
 			throw new Exception(__('Erreur données de localisation invalide ou vide : ', __FILE__) . json_encode($result));
 		}
@@ -66,13 +70,13 @@ class gsl extends eqLogic {
 	}
 
 	public static function google_locationData() {
-		if (!file_exists(jeedom::getTmpFolder('gsl') . '/cookies.txt')) {
-			self::google_connect();
+		if (!file_exists(self::$_cookiePath)) {
+			log::add('gsl', 'error', __('Cookie absent, veuillez consulter la documentation pour configurer le plugin.', __FILE__));
 		}
 		try {
 			$result = self::google_callLocationUrl();
 		} catch (Exception $e) {
-			self::google_connect();
+			//self::google_connect();
 			$result = self::google_callLocationUrl();
 		}
 		$result = $result[0];
@@ -85,155 +89,21 @@ class gsl extends eqLogic {
 				'address' => $user[1][4],
 				'timestamp' => $user[1][2],
 				'coordinated' => $user[1][1][2] . ',' . $user[1][1][1],
-				'battery' => $user[13][1]
+				'battery' => $user[13][1],
+				'charging' => $user[13][0],
+              	'accuracy' => $user[1][3]
 			);
 		}
 		return $return;
 	}
 
 	public static function google_logout() {
-		if (!file_exists(jeedom::getTmpFolder('gsl') . '/cookies.txt')) {
+		if (!file_exists(self::$_cookiePath)) {
 			return;
 		}
-		unlink(jeedom::getTmpFolder('gsl') . '/cookies.txt');
+		unlink(self::$_cookiePath);
 	}
 
-	public static function google_connect() {
-		self::google_logout();
-		$data = array();
-		/*************************STAGE 1*******************************/
-		log::add('gsl', 'debug', __('Stage 1 : Connection à google', __FILE__));
-		$ch = curl_init('https://accounts.google.com/ServiceLogin?rip=1&nojavascript=1');
-		curl_setopt($ch, CURLOPT_COOKIEJAR, jeedom::getTmpFolder('gsl') . '/cookies.txt');
-		curl_setopt($ch, CURLOPT_COOKIEFILE, jeedom::getTmpFolder('gsl') . '/cookies.txt');
-		curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-		curl_setopt($ch, CURLOPT_TIMEOUT, 3);
-		curl_setopt($ch, CURLOPT_HEADER, 1);
-		$response = curl_exec($ch);
-		$info = curl_getinfo($ch);
-		log::add('gsl', 'debug', __('Stage 1 : Connection réussie, reponse : ', __FILE__) . $info['http_code']);
-		if (!empty($info['http_code']) && $info['http_code'] == 302) {
-			return true;
-		}
-		if (empty($info['http_code']) || $info['http_code'] != 200) {
-			throw new Exception(__('Erreur stage 1 : code retour invalide : ', __FILE__) . $info['http_code']);
-		}
-		curl_close($ch);
-		$headers = self::get_headers_from_curl_response($response);
-		if (!isset($headers['Set-Cookie'])) {
-			throw new Exception(__('Erreur stage 1 : aucun cookie', __FILE__));
-		}
-		$data['cookie'] = array('google.com' => self::processCookies($headers['Set-Cookie']));
-		preg_match_all('/<input type="hidden" name="gxf" value="(.*?)">/m', $response, $matches);
-		if (!isset($matches[1]) || !isset($matches[1][0])) {
-			throw new Exception(__('Erreur stage 1 : champs gfx non trouvé', __FILE__));
-		}
-		$data['gfx'] = $matches[1][0];
-		log::add('gsl', 'debug', __('Stage 1 : Connection réussie, sauvegarde du cookie', __FILE__));
-
-		/*************************STAGE 2*******************************/
-		log::add('gsl', 'debug', __('Stage 2 : envoi du mail', __FILE__));
-		$ch = curl_init("https://accounts.google.com/signin/v1/lookup");
-		curl_setopt($ch, CURLOPT_COOKIEJAR, jeedom::getTmpFolder('gsl') . '/cookies.txt');
-		curl_setopt($ch, CURLOPT_COOKIEFILE, jeedom::getTmpFolder('gsl') . '/cookies.txt');
-		curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-		curl_setopt($ch, CURLOPT_TIMEOUT, 3);
-		curl_setopt($ch, CURLOPT_HEADER, 1);
-		curl_setopt($ch, CURLOPT_HTTPHEADER, array('Cookie' => 'GAPS=' . $data['cookie']['google.com']['GAPS']));
-		curl_setopt($ch, CURLOPT_POST, true);
-		curl_setopt($ch, CURLOPT_POSTFIELDS, array(
-			'Page' => 'PasswordSeparationSignIn',
-			'gxf' => $data['gfx'],
-			'rip' => '1',
-			'ProfileInformation' => '',
-			'SessionState' => '',
-			'bgresponse' => 'js_disabled',
-			'pstMsg' => '0',
-			'checkConnection' => '',
-			'checkedDomains' => 'youtube',
-			'Email' => config::byKey('google_user', 'gsl'),
-			'identifiertoken' => '',
-			'identifiertoken_audio' => '',
-			'identifier-captcha-input' => '',
-			'signIn' => 'Weiter',
-			'Passwd' => '',
-			'PersistentCookie' => 'yes',
-		));
-		$response = curl_exec($ch);
-		$info = curl_getinfo($ch);
-		log::add('gsl', 'debug', __('Stage 2 : Connection réussie, reponse : ', __FILE__) . $info['http_code']);
-		if (empty($info['http_code']) || $info['http_code'] != 200) {
-			throw new Exception(__('Erreur stage 2 : code retour invalide : ', __FILE__) . $info['http_code']);
-		}
-		$headers = self::get_headers_from_curl_response($response);
-		if (!isset($headers['Set-Cookie'])) {
-			throw new Exception(__('Erreur stage 2 : aucun cookie', __FILE__));
-		}
-		$data['cookie'] = array('google.com' => array_merge($data['cookie']['google.com'], self::processCookies($headers['Set-Cookie'], $data['cookie'])));
-
-		preg_match_all('/<input id="profile-information" name="ProfileInformation" type="hidden" value="(.*?)">/m', $response, $matches);
-		if (!isset($matches[1]) || !isset($matches[1][0])) {
-			throw new Exception(__('Erreur stage 2 : champs ProfileInformation non trouvé', __FILE__));
-		}
-		$data['ProfileInformation'] = $matches[1][0];
-
-		preg_match_all('/<input id="session-state" name="SessionState" type="hidden" value="(.*?)">/m', $response, $matches);
-		if (!isset($matches[1]) || !isset($matches[1][0])) {
-			throw new Exception(__('Erreur stage 2 : champs SessionState non trouvé', __FILE__));
-		}
-		$data['SessionState'] = $matches[1][0];
-		log::add('gsl', 'debug', __('Stage 2 : Connection réussie, sauvegarde du cookie', __FILE__));
-
-		/*************************STAGE 3*******************************/
-		log::add('gsl', 'debug', __('Stage 3 : envoi du mot de passe', __FILE__));
-
-		$ch = curl_init("https://accounts.google.com/signin/challenge/sl/password");
-		curl_setopt($ch, CURLOPT_COOKIEJAR, jeedom::getTmpFolder('gsl') . '/cookies.txt');
-		curl_setopt($ch, CURLOPT_COOKIEFILE, jeedom::getTmpFolder('gsl') . '/cookies.txt');
-		curl_setopt($ch, CURLOPT_RETURNTRANSFER, 1);
-		curl_setopt($ch, CURLOPT_TIMEOUT, 3);
-		curl_setopt($ch, CURLOPT_HEADER, 1);
-		curl_setopt($ch, CURLOPT_HTTPHEADER, array(
-			'Cookie' => 'GAPS=' . $data['cookie']['google.com']['GAPS'] . ';GALX=' . $data['cookie']['google.com']['GALX'],
-			'Origin' => 'https://accounts.google.com',
-			'Referer' => 'https://accounts.google.com/signin/v1/lookup',
-			'Upgrade-Insecure-Requests' => '1',
-		));
-		curl_setopt($ch, CURLOPT_POST, true);
-		curl_setopt($ch, CURLOPT_POSTFIELDS, array(
-			'Page' => 'PasswordSeparationSignIn',
-			'GALX' => $data['cookie']['google.com']['GALX'],
-			'gxf' => $data['gfx'],
-			'checkedDomains' => 'youtube',
-			'pstMsg' => '0',
-			'rip' => '1',
-			'ProfileInformation' => $data['ProfileInformation'],
-			'SessionState' => $data['SessionState'],
-			'_utf8' => '☃',
-			'bgresponse' => 'js_disabled',
-			'checkConnection' => '',
-			'Email' => config::byKey('google_user', 'gsl'),
-			'signIn' => 'Weiter',
-			'Passwd' => config::byKey('google_password', 'gsl'),
-			'PersistentCookie' => 'yes',
-			'rmShown' => '1',
-		));
-		$response = curl_exec($ch);
-		$info = curl_getinfo($ch);
-		log::add('gsl', 'debug', 'Stage 3 : Connection réussie, reponse : ' . $info['http_code']);
-		if (empty($info['http_code']) || $info['http_code'] != 302) {
-			throw new Exception(__('Erreur stage 3 : connection etablie mais echec de l\'autentification, code 302 attendu : ', __FILE__) . $info['http_code']);
-		}
-		$headers = self::get_headers_from_curl_response($response);
-		if (!isset($headers['Set-Cookie'])) {
-			throw new Exception(__('Erreur stage 3 : aucun cookie', __FILE__));
-		}
-		$data['cookie'] = array('google.com' => array_merge($data['cookie']['google.com'], self::processCookies($headers['Set-Cookie'], $data['cookie'])));
-		if (!isset($headers['Location'])) {
-			throw new Exception(__('Erreur stage 3 : aucun adresse de redirection', __FILE__));
-		}
-		return true;
-	}
 
 	public static function processCookies($_cookie) {
 		return array(explode('=', explode(';', $_cookie)[0])[0] => explode('=', explode(';', $_cookie)[0])[1]);
@@ -253,6 +123,12 @@ class gsl extends eqLogic {
 		}
 		return $headers;
 	}
+  
+  public static function google_saveCookie($_cookie){
+		$fp = fopen(self::$_cookiePath, 'w');
+		fwrite($fp, $_cookie);
+		fclose($fp);
+  }
 
 	public static function pull($_force = false) {
 		if (!$_force) {
@@ -286,6 +162,8 @@ class gsl extends eqLogic {
 			$changed = $eqLogic->checkAndUpdateCmd('image', $location['image']) || $changed;
 			$changed = $eqLogic->checkAndUpdateCmd('address', $location['address'], $timestamp) || $changed;
 			$changed = $eqLogic->checkAndUpdateCmd('battery', $location['battery'], $timestamp) || $changed;
+			$changed = $eqLogic->checkAndUpdateCmd('charging', $location['charging'], $timestamp) || $changed;
+			$changed = $eqLogic->checkAndUpdateCmd('accuracy', $location['accuracy'], $timestamp) || $changed;
 			$cmdgeoloc = $eqLogic->getConfiguration('cmdgeoloc', null);
 			if ($cmdgeoloc !== null) {
 				$cmdUpdate = cmd::byId(str_replace('#', '', $cmdgeoloc));
@@ -418,6 +296,26 @@ class gsl extends eqLogic {
 				$cmd->setSubType('string');
 				$cmd->save();
 			}
+			$cmd = $this->getCmd(null, 'charging');
+			if (!is_object($cmd)) {
+				$cmd = new cmd();
+				$cmd->setName('charge');
+				$cmd->setEqLogic_id($this->getId());
+				$cmd->setLogicalId('charging');
+				$cmd->setType('info');
+				$cmd->setSubType('binary');
+				$cmd->save();
+			}
+			$cmd = $this->getCmd(null, 'accuracy');
+			if (!is_object($cmd)) {
+				$cmd = new cmd();
+				$cmd->setName('precision');
+				$cmd->setEqLogic_id($this->getId());
+				$cmd->setLogicalId('accuracy');
+				$cmd->setType('info');
+				$cmd->setSubType('string');
+				$cmd->save();
+			}
 		}
 	}
 
@@ -524,12 +422,13 @@ class gsl extends eqLogic {
 				$replace['#adresses#'] .= '<div class="gsl-address" id="gsl-address-' . $this->getLogicalId() . '-' . $eqLogic->getId() . '">';
 				$replace['#adresses#'] .= '<span class="pull-right" style="text-align: center;"><img style="border: 2px solid white; background-color:' . $color . ';cursor:pointer; margin-top:5px;width:50px; height:50px;border-radius: 50% !important;" src="' . $data[$eqLogic->getId()]['image'] . '" />';
         if(isset($data[$eqLogic->getId()]['battery']) && $data[$eqLogic->getId()]['battery'] != '') {
-            $replace['#adresses#'] .= '<br/><span style="font-size:0.7em;"><i class="fa ' . $data[$eqLogic->getId()]['battery_icon'] . '"></i> ' . $data[$eqLogic->getId()]['battery'] . '%</span>';
+            $replace['#adresses#'] .= '<br/><span style="font-size:0.7em;">'.($data[$eqLogic->getId()]['charging'] ? '<i class="fas fa-bolt"></i> ' : '' ).'<i class="fa ' . $data[$eqLogic->getId()]['battery_icon'] . '"></i> ' . $data[$eqLogic->getId()]['battery'] . '%</span>';
         }
         $replace['#adresses#'] .= '</span>';
 				$replace['#adresses#'] .= '<span style="font-size:0.8em;">' . $data[$eqLogic->getId()]['name'] . '</span><br/>';
 				$replace['#adresses#'] .= '<span>' . $data[$eqLogic->getId()]['address'] . '</span><br/>';
-				$replace['#adresses#'] .= '<span style="font-size:0.7em;">' . $data[$eqLogic->getId()]['horodatage'] . '</span>';
+				$replace['#adresses#'] .= '<span style="font-size:0.7em;">' . $data[$eqLogic->getId()]['horodatage'] . '</span><br/>';
+				$replace['#adresses#'] .= '<span style="font-size:0.7em;">Précision : ' . $data[$eqLogic->getId()]['accuracy'] . 'm</span>';
 				$replace['#adresses#'] .= '</div>';
 				$replace['#adresses#'] .= '<hr/>';
 			}
@@ -545,9 +444,10 @@ class gsl extends eqLogic {
 			$data[$this->getId()]['color'] = $color;
 			$replace['#adresses#'] = '<span>' . $data[$this->getId()]['address'] . '</span><br/>';
 			if(isset($data[$this->getId()]['battery']) && $data[$this->getId()]['battery'] != '') {
-                $replace['#adresses#'] .= '<span style="font-size:0.7em;"><i class="fa ' . $data[$this->getId()]['battery_icon'] . '"></i> ' . $data[$this->getId()]['battery'] . '%</span> - ';
+                $replace['#adresses#'] .= '<span style="font-size:0.7em;">'.($data[$this->getId()]['charging'] ? '<i class="fas fa-bolt"></i> ' : '' ).'<i class="fa ' . $data[$this->getId()]['battery_icon'] . '"></i> ' . $data[$this->getId()]['battery'] . '%</span> - ';
             }
-			$replace['#adresses#'] .= '<span style="font-size:0.7em;">' . $data[$this->getId()]['horodatage'] . '</span>';
+			$replace['#adresses#'] .= '<span style="font-size:0.7em;">' . $data[$this->getId()]['horodatage'] . '</span><br/>';
+				$replace['#adresses#'] .= '<span style="font-size:0.7em;">Précision : ' . $data[$this->getId()]['accuracy'] . 'm</span>';
 			$replace['#json#'] = str_replace("'", "\'", json_encode($data));
 			$replace['#height-map#'] = ($version == 'dashboard') ? $replace['#height#'] - 100 : 170;
 			return $this->postToHtml($_version, template_replace($replace, getTemplate('core', $version, 'gsl', 'gsl')));
@@ -591,6 +491,13 @@ class gsl extends eqLogic {
 		}
 		return $return;
 	}
+  
+	public function postUpdate() {
+		if (file_exists(jeedom::getTmpFolder('gsl') . '/cookies.txt') && !file_exists(self::$_cookiePath)) {
+          	copy(jeedom::getTmpFolder('gsl') . '/cookies.txt', self::$_cookiePath);
+          	unlink(jeedom::getTmpFolder('gsl') . '/cookies.txt');
+		}
+   	}
 
 	/*     * **********************Getteur Setteur*************************** */
 }